fix(aria): make sure snapshot diff works with iframes
