+++ packages/playwright-core/src/utils/isomorphic/ariaSnapshot.ts
@@ -14,6 +14,9 @@
  * limitations under the License.
  */
 
+import { escapeRegExp, longestCommonSubstring } from './stringUtils';
+import { yamlEscapeKeyIfNeeded, yamlEscapeValueIfNeeded } from './yaml';
+
 // https://www.w3.org/TR/wai-aria-1.2/#role_definitions
 
 export type AriaRole = 'alert' | 'alertdialog' | 'application' | 'article' | 'banner' | 'blockquote' | 'button' | 'caption' | 'cell' | 'checkbox' | 'code' | 'columnheader' | 'combobox' |
@@ -24,7 +27,9 @@ export type AriaRole = 'alert' | 'alertdialog' | 'application' | 'article' | 'ba
   'spinbutton' | 'status' | 'strong' | 'subscript' | 'superscript' | 'switch' | 'tab' | 'table' | 'tablist' | 'tabpanel' | 'term' | 'textbox' | 'time' | 'timer' |
   'toolbar' | 'tooltip' | 'tree' | 'treegrid' | 'treeitem';
 
-export type AriaProps = {
+export type AriaTreeMode = 'ai' | 'expect' | 'codegen' | 'autoexpect';
+
+type AriaProps = {
   checked?: boolean | 'mixed';
   disabled?: boolean;
   expanded?: boolean;
@@ -32,6 +37,16 @@ export type AriaProps = {
   level?: number;
   pressed?: boolean | 'mixed';
   selected?: boolean;
+};
+
+export type SerializableAriaNode = AriaProps & {
+  role: AriaRole | 'fragment' | 'iframe';
+  name: string;
+  ref?: string;
+  children: (SerializableAriaNode | string)[];
+  box: { visible: boolean, inline: boolean, cursor?: string };
+  receivesPointerEvents: boolean;
+  props: Record<string, string>;
 };
 
 // We pass parsed template between worlds using JSON, make it easy.
@@ -491,4 +506,220 @@ export class ParserError extends Error {
     super(message);
     this.pos = pos;
   }
+}
+
+function buildByRefMap(root: SerializableAriaNode | undefined, map: Map<string, SerializableAriaNode> = new Map()): Map<string, SerializableAriaNode> {
+  if (root?.ref)
+    map.set(root.ref, root);
+  for (const child of root?.children || []) {
+    if (typeof child !== 'string')
+      buildByRefMap(child, map);
+  }
+  return map;
+}
+
+function arePropsEqual(a: SerializableAriaNode, b: SerializableAriaNode): boolean {
+  const aKeys = Object.keys(a.props);
+  const bKeys = Object.keys(b.props);
+  return aKeys.length === bKeys.length && aKeys.every(k => a.props[k] === b.props[k]);
+}
+
+type RenderOptions = {
+  renderCursorPointer?: boolean,
+  renderActive?: boolean,
+  renderStringsAsRegex?: boolean,
+};
+
+function toRenderOptions(mode: AriaTreeMode): RenderOptions {
+  if (mode === 'ai')
+    return { renderActive: true, renderCursorPointer: true };
+  if (mode === 'codegen')
+    return { renderStringsAsRegex: true };
+  return {};
+}
+
+export function renderAriaTree(root: SerializableAriaNode, mode: AriaTreeMode, previousRoot?: SerializableAriaNode): string {
+  const options = toRenderOptions(mode);
+  const lines: string[] = [];
+  const includeText = options.renderStringsAsRegex ? textContributesInfo : () => true;
+  const renderString = options.renderStringsAsRegex ? convertToBestGuessRegex : (str: string) => str;
+  const previousByRef = buildByRefMap(previousRoot);
+
+  const visitText = (text: string, indent: string) => {
+    const escaped = yamlEscapeValueIfNeeded(renderString(text));
+    if (escaped)
+      lines.push(indent + '- text: ' + escaped);
+  };
+
+  const createKey = (ariaNode: SerializableAriaNode, renderCursorPointer: boolean): string => {
+    let key = ariaNode.role;
+    // Yaml has a limit of 1024 characters per key, and we leave some space for role and attributes.
+    if (ariaNode.name && ariaNode.name.length <= 900) {
+      const name = renderString(ariaNode.name);
+      if (name) {
+        const stringifiedName = name.startsWith('/') && name.endsWith('/') ? name : JSON.stringify(name);
+        key += ' ' + stringifiedName;
+      }
+    }
+    if (ariaNode.checked === 'mixed')
+      key += ` [checked=mixed]`;
+    if (ariaNode.checked === true)
+      key += ` [checked]`;
+    if (ariaNode.disabled)
+      key += ` [disabled]`;
+    if (ariaNode.expanded)
+      key += ` [expanded]`;
+    if (ariaNode.active && options.renderActive)
+      key += ` [active]`;
+    if (ariaNode.level)
+      key += ` [level=${ariaNode.level}]`;
+    if (ariaNode.pressed === 'mixed')
+      key += ` [pressed=mixed]`;
+    if (ariaNode.pressed === true)
+      key += ` [pressed]`;
+    if (ariaNode.selected === true)
+      key += ` [selected]`;
+
+    if (ariaNode.ref) {
+      key += ` [ref=${ariaNode.ref}]`;
+      if (renderCursorPointer && hasPointerCursor(ariaNode))
+        key += ' [cursor=pointer]';
+    }
+    return key;
+  };
+
+  const getSingleInlinedTextChild = (ariaNode: SerializableAriaNode | undefined): string | undefined => {
+    return ariaNode?.children.length === 1 && typeof ariaNode.children[0] === 'string' && !Object.keys(ariaNode.props).length ? ariaNode.children[0] : undefined;
+  };
+
+  const visit = (ariaNode: SerializableAriaNode, indent: string, renderCursorPointer: boolean, previousNode: SerializableAriaNode | undefined): { unchanged: boolean } => {
+    if (ariaNode.ref)
+      previousNode = previousByRef.get(ariaNode.ref);
+
+    const linesBefore = lines.length;
+    const key = createKey(ariaNode, renderCursorPointer);
+    const escapedKey = indent + '- ' + yamlEscapeKeyIfNeeded(key);
+    const inCursorPointer = renderCursorPointer && !!ariaNode.ref && hasPointerCursor(ariaNode);
+    const singleInlinedTextChild = getSingleInlinedTextChild(ariaNode);
+
+    // Whether ariaNode's subtree is the same as previousNode's, and can be replaced with just a ref.
+    let unchanged = !!previousNode && key === createKey(previousNode, renderCursorPointer) && arePropsEqual(ariaNode, previousNode);
+
+    if (!ariaNode.children.length && !Object.keys(ariaNode.props).length) {
+      // Leaf node without children.
+      lines.push(escapedKey);
+    } else if (singleInlinedTextChild !== undefined) {
+      // Leaf node with just some text inside.
+      // Unchanged when the previous node also had the same single text child.
+      unchanged = unchanged && getSingleInlinedTextChild(previousNode) === singleInlinedTextChild;
+
+      const shouldInclude = includeText(ariaNode, singleInlinedTextChild);
+      if (shouldInclude)
+        lines.push(escapedKey + ': ' + yamlEscapeValueIfNeeded(renderString(singleInlinedTextChild)));
+      else
+        lines.push(escapedKey);
+    } else {
+      // Node with (optional) props and some children.
+      lines.push(escapedKey + ':');
+      for (const [name, value] of Object.entries(ariaNode.props))
+        lines.push(indent + '  - /' + name + ': ' + yamlEscapeValueIfNeeded(value));
+
+      // All children must be the same.
+      unchanged = unchanged && previousNode?.children.length === ariaNode.children.length;
+
+      const childIndent = indent + '  ';
+      for (let childIndex = 0 ; childIndex < ariaNode.children.length; childIndex++) {
+        const child = ariaNode.children[childIndex];
+        if (typeof child === 'string') {
+          unchanged = unchanged && previousNode?.children[childIndex] === child;
+          if (includeText(ariaNode, child))
+            visitText(child, childIndent);
+        } else {
+          const previousChild = previousNode?.children[childIndex];
+          const childResult = visit(child, childIndent, renderCursorPointer && !inCursorPointer, typeof previousChild !== 'string' ? previousChild : undefined);
+          unchanged = unchanged && childResult.unchanged;
+        }
+      }
+    }
+
+    if (unchanged && ariaNode.ref) {
+      // Replace the whole subtree with a single reference.
+      lines.splice(linesBefore);
+      lines.push(indent + `- ref=${ariaNode.ref} [unchanged]`);
+    }
+
+    return { unchanged };
+  };
+
+  // Do not render the root fragment, just its children.
+  const nodesToRender = root.role === 'fragment' ? root.children : [root];
+  for (const nodeToRender of nodesToRender) {
+    if (typeof nodeToRender === 'string')
+      visitText(nodeToRender, '');
+    else
+      visit(nodeToRender, '', !!options.renderCursorPointer, undefined);
+  }
+  return lines.join('\n');
+}
+
+function convertToBestGuessRegex(text: string): string {
+  const dynamicContent = [
+    // 2mb
+    { regex: /\b[\d,.]+[bkmBKM]+\b/, replacement: '[\\d,.]+[bkmBKM]+' },
+    // 2ms, 20s
+    { regex: /\b\d+[hmsp]+\b/, replacement: '\\d+[hmsp]+' },
+    { regex: /\b[\d,.]+[hmsp]+\b/, replacement: '[\\d,.]+[hmsp]+' },
+    // Do not replace single digits with regex by default.
+    // 2+ digits: [Issue 22, 22.3, 2.33, 2,333]
+    { regex: /\b\d+,\d+\b/, replacement: '\\d+,\\d+' },
+    { regex: /\b\d+\.\d{2,}\b/, replacement: '\\d+\\.\\d+' },
+    { regex: /\b\d{2,}\.\d+\b/, replacement: '\\d+\\.\\d+' },
+    { regex: /\b\d{2,}\b/, replacement: '\\d+' },
+  ];
+
+  let pattern = '';
+  let lastIndex = 0;
+
+  const combinedRegex = new RegExp(dynamicContent.map(r => '(' + r.regex.source + ')').join('|'), 'g');
+  text.replace(combinedRegex, (match, ...args) => {
+    const offset = args[args.length - 2];
+    const groups = args.slice(0, -2);
+    pattern += escapeRegExp(text.slice(lastIndex, offset));
+    for (let i = 0; i < groups.length; i++) {
+      if (groups[i]) {
+        const { replacement } = dynamicContent[i];
+        pattern += replacement;
+        break;
+      }
+    }
+    lastIndex = offset + match.length;
+    return match;
+  });
+  if (!pattern)
+    return text;
+
+  pattern += escapeRegExp(text.slice(lastIndex));
+  return String(new RegExp(pattern));
+}
+
+function textContributesInfo(node: SerializableAriaNode, text: string): boolean {
+  if (!text.length)
+    return false;
+
+  if (!node.name)
+    return true;
+
+  if (node.name.length > text.length)
+    return false;
+
+  // Figure out if text adds any value. "longestCommonSubstring" is expensive, so limit strings length.
+  const substr = (text.length <= 200 && node.name.length <= 200) ? longestCommonSubstring(text, node.name) : '';
+  let filtered = text;
+  while (substr && filtered.includes(substr))
+    filtered = filtered.replace(substr, '');
+  return filtered.trim().length / text.length > 0.1;
+}
+
+function hasPointerCursor(ariaNode: SerializableAriaNode): boolean {
+  return ariaNode.box.cursor === 'pointer';
 }
