+++ tests/page/page-aria-snapshot-ai.spec.ts
@@ -503,26 +503,26 @@ it('should create incremental snapshots on multiple tracks', async ({ page }) =>
 it('should create incremental snapshot for attribute change', async ({ page }) => {
   await page.setContent(`<button>a button</button>`);
   await page.evaluate(() => document.querySelector('button').focus());
-  expect(await snapshotForAI(page, { mode: 'incremental' })).toContainYaml(`
+  expect(await snapshotForAI(page, { track: 'track', mode: 'incremental' })).toContainYaml(`
     - button "a button" [active] [ref=e2]
   `);
 
   await page.evaluate(() => document.querySelector('button').blur());
-  expect(await snapshotForAI(page, { mode: 'incremental' })).toContainYaml(`
+  expect(await snapshotForAI(page, { track: 'track', mode: 'incremental' })).toContainYaml(`
     - button "a button" [ref=e2]
   `);
 });
 
 it('should create incremental snapshot for child removal', async ({ page }) => {
   await page.setContent(`<li><button>a button</button><span>some text</span></li>`);
-  expect(await snapshotForAI(page, { mode: 'incremental' })).toContainYaml(`
+  expect(await snapshotForAI(page, { track: 'track', mode: 'incremental' })).toContainYaml(`
     - listitem [ref=e2]:
       - button "a button" [ref=e3]
       - text: some text
   `);
 
   await page.evaluate(() => document.querySelector('span').remove());
-  expect(await snapshotForAI(page, { mode: 'incremental' })).toContainYaml(`
+  expect(await snapshotForAI(page, { track: 'track', mode: 'incremental' })).toContainYaml(`
     - listitem [ref=e2]:
       - ref=e3 [unchanged]
   `);
@@ -530,13 +530,13 @@ it('should create incremental snapshot for child removal', async ({ page }) => {
 
 it('should create incremental snapshot for child addition', async ({ page }) => {
   await page.setContent(`<li><button>a button</button><span style="display:none">some text</span></li>`);
-  expect(await snapshotForAI(page, { mode: 'incremental' })).toContainYaml(`
+  expect(await snapshotForAI(page, { track: 'track', mode: 'incremental' })).toContainYaml(`
     - listitem [ref=e2]:
       - button "a button" [ref=e3]
   `);
 
   await page.evaluate(() => document.querySelector('span').style.display = 'inline');
-  expect(await snapshotForAI(page, { mode: 'incremental' })).toContainYaml(`
+  expect(await snapshotForAI(page, { track: 'track', mode: 'incremental' })).toContainYaml(`
     - listitem [ref=e2]:
       - ref=e3 [unchanged]
       - text: some text
@@ -545,13 +545,13 @@ it('should create incremental snapshot for child addition', async ({ page }) => 
 
 it('should create incremental snapshot for prop change', async ({ page }) => {
   await page.setContent(`<a href="about:blank" style="cursor:pointer">a link</a>`);
-  expect(await snapshotForAI(page, { mode: 'incremental' })).toContainYaml(`
+  expect(await snapshotForAI(page, { track: 'track', mode: 'incremental' })).toContainYaml(`
     - link "a link" [ref=e2] [cursor=pointer]:
       - /url: about:blank
   `);
 
   await page.evaluate(() => document.querySelector('a').setAttribute('href', 'https://playwright.dev'));
-  expect(await snapshotForAI(page, { mode: 'incremental' })).toContainYaml(`
+  expect(await snapshotForAI(page, { track: 'track', mode: 'incremental' })).toContainYaml(`
     - link "a link" [ref=e2] [cursor=pointer]:
       - /url: https://playwright.dev
   `);
@@ -559,13 +559,13 @@ it('should create incremental snapshot for prop change', async ({ page }) => {
 
 it('should create incremental snapshot for cursor change', async ({ page }) => {
   await page.setContent(`<a href="about:blank" style="cursor:pointer">a link</a>`);
-  expect(await snapshotForAI(page, { mode: 'incremental' })).toContainYaml(`
+  expect(await snapshotForAI(page, { track: 'track', mode: 'incremental' })).toContainYaml(`
     - link "a link" [ref=e2] [cursor=pointer]:
       - /url: about:blank
   `);
 
   await page.evaluate(() => document.querySelector('a').style.cursor = 'default');
-  expect(await snapshotForAI(page, { mode: 'incremental' })).toContainYaml(`
+  expect(await snapshotForAI(page, { track: 'track', mode: 'incremental' })).toContainYaml(`
     - link "a link" [ref=e2]:
       - /url: about:blank
   `);
@@ -573,24 +573,60 @@ it('should create incremental snapshot for cursor change', async ({ page }) => {
 
 it('should create incremental snapshot for name change', async ({ page }) => {
   await page.setContent(`<button><span>a button</span></button>`);
-  expect(await snapshotForAI(page, { mode: 'incremental' })).toContainYaml(`
+  expect(await snapshotForAI(page, { track: 'track', mode: 'incremental' })).toContainYaml(`
     - button "a button" [ref=e2]
   `);
 
   await page.evaluate(() => document.querySelector('span').textContent = 'new button');
-  expect(await snapshotForAI(page, { mode: 'incremental' })).toContainYaml(`
+  expect(await snapshotForAI(page, { track: 'track', mode: 'incremental' })).toContainYaml(`
     - button "new button" [ref=e3]
   `);
 });
 
 it('should create incremental snapshot for text change', async ({ page }) => {
   await page.setContent(`<li><span>an item</span></li>`);
-  expect(await snapshotForAI(page, { mode: 'incremental' })).toContainYaml(`
+  expect(await snapshotForAI(page, { track: 'track', mode: 'incremental' })).toContainYaml(`
     - listitem [ref=e2]: an item
   `);
 
   await page.evaluate(() => document.querySelector('span').textContent = 'new text');
-  expect(await snapshotForAI(page, { mode: 'incremental' })).toContainYaml(`
+  expect(await snapshotForAI(page, { track: 'track', mode: 'incremental' })).toContainYaml(`
     - listitem [ref=e2]: new text
+  `);
+});
+
+it('should not create incremental snapshots without tracks', async ({ page }) => {
+  await page.setContent(`<ul><li><button>a button</button></li><li><span>a span</span></li><li id=hidden-li style="display:none">some text</li></ul>`);
+
+  expect(await snapshotForAI(page, { mode: 'incremental' })).toContainYaml(`
+    - list [ref=e2]:
+      - listitem [ref=e3]:
+        - button "a button" [ref=e4]
+      - listitem [ref=e5]: a span
+  `);
+  expect(await snapshotForAI(page, { mode: 'incremental' })).toContainYaml(`
+    - list [ref=e2]:
+      - listitem [ref=e3]:
+        - button "a button" [ref=e4]
+      - listitem [ref=e5]: a span
+  `);
+});
+
+it('should create incremental snapshot in an iframe', async ({ page }) => {
+  await page.setContent(`<iframe srcdoc="
+    <li><button>a button</button><span style='display:none'>some text</span></li>
+  "></iframe>`);
+  expect(await snapshotForAI(page, { track: 'track', mode: 'incremental' })).toContainYaml(`
+    - iframe [ref=e2]:
+      - listitem [ref=f1e2]:
+        - button "a button" [ref=f1e3]
+  `);
+
+  await page.frames()[1].evaluate(() => document.querySelector('span').style.display = 'inline');
+  expect(await snapshotForAI(page, { track: 'track', mode: 'incremental' })).toContainYaml(`
+    - iframe [ref=e2]:
+      - listitem [ref=f1e2]:
+        - ref=f1e3 [unchanged]
+        - text: some text
   `);
 });
